+++ src/index.ts
@@ -1,3 +1,4 @@
 export { EVENTS } from "./const.js";
 export { Queue } from "./Queue.js";
 export type { TaskFn, FilterFn, QueueOptions, QueueDestroyInfo } from "./types.js";
+export type { Task } from "./Task.js";